requirements.txt
Django>=5.0,<6.0
strawberry-graphql[django]>=0.236,<1.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
PyJWT>=2.9,<3.0
python-dotenv>=1.0,<2.0
